locust_files/requirements.txt
pyaml
p4python==2017.2.1615960
numpy
pyzmq
locustio
mimesis
